before ip_address -> service_ip / host_ip refactor
