+++ src/Makefile
@@ -1,0 +1,7 @@
+.PHONY: help
+help:  ## You're looking at it
+	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | sort | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-10s\033[0m %s\n", $$1, $$2}'
+
+.PHONY: test
+test:  ## Run unit tests
+	tox -e xenial
