+++ src/tests/configure_nerve_test.py
@@ -1,10 +1,19 @@
 import copy
 import mock
+from mock import call
+from mock import patch
+from mock import mock_open
+from mock import MagicMock
+from mock import Mock
 import pytest
 import sys
 import multiprocessing
-from nerve_tools import configure_nerve
 from contextlib import contextmanager
+
+from typing import List
+
+from nerve_tools import configure_nerve
+from nerve_tools.configure_nerve import generate_subconfiguration
 
 
 try:
@@ -14,11 +23,11 @@ except NotImplementedError:
 
 
 def test_get_named_zookeeper_topology():
-    m = mock.mock_open()
-    with mock.patch(
+    m = mock_open()
+    with patch(
         'nerve_tools.configure_nerve.open',
         m, create=True
-    ), mock.patch(
+    ), patch(
         'yaml.load', return_value=[['foo', 42]]
     ):
         zk_topology = configure_nerve.get_named_zookeeper_topology(
@@ -37,7 +46,7 @@ def get_labels_by_service_and_port(service, port, labels_dir):
         return {}
 
 
-def get_current_location(typ):
+def get_current_location(typ: str) -> str:
     return {
         'ecosystem': 'my_ecosystem',
         'superregion': 'my_superregion',
@@ -46,7 +55,7 @@ def get_current_location(typ):
     }[typ]
 
 
-def convert_location_type(src_loc, src_typ, dst_typ):
+def convert_location_type(src_loc: str, src_typ: str, dst_typ: str) -> List[str]:
     if src_typ == dst_typ:
         return [src_loc]
     return {
@@ -126,15 +135,15 @@ def expected_sub_config():
 
 
 @pytest.fixture
-def expected_sub_config_with_envoy_listeners(expected_sub_config):
+def expected_sub_config_with_envoy_ingress_listeners(expected_sub_config):
     expected_sub_config.update({
-        'test_service.my_superregion:ip_address.1234': {
+        'test_service.my_superregion:service_ip.1234': {
             'zk_hosts': ['1.2.3.4', '2.3.4.5'],
             'zk_path': '/envoy/global/test_service',
             'checks': [{
                 'rise': 1,
                 'uri': '/https/test_service/35000/status',
-                'host': '127.0.0.1',
+                'host': 'host_ip',
                 'timeout': 2.0,
                 'open_timeout': 2.0,
                 'fall': 2,
@@ -142,7 +151,7 @@ def expected_sub_config_with_envoy_listeners(expected_sub_config):
                 'port': 6666,
                 'headers': {'Host': 'test_service'},
             }],
-            'host': 'ip_address',
+            'host': 'host_ip',
             'check_interval': 3.0,
             'port': 35000,
             'weight': mock.sentinel.weight,
@@ -155,13 +164,13 @@ def expected_sub_config_with_envoy_listeners(expected_sub_config):
                 'paasta_instance': 'canary',
             },
         },
-        'test_service.another_superregion:ip_address.1234': {
+        'test_service.another_superregion:service_ip.1234': {
             'zk_hosts': ['3.4.5.6', '4.5.6.7'],
             'zk_path': '/envoy/global/test_service',
             'checks': [{
                 'rise': 1,
                 'uri': '/https/test_service/35000/status',
-                'host': '127.0.0.1',
+                'host': 'host_ip',
                 'timeout': 2.0,
                 'open_timeout': 2.0,
                 'fall': 2,
@@ -169,7 +178,7 @@ def expected_sub_config_with_envoy_listeners(expected_sub_config):
                 'port': 6666,
                 'headers': {'Host': 'test_service'},
             }],
-            'host': 'ip_address',
+            'host': 'host_ip',
             'check_interval': 3.0,
             'port': 35000,
             'weight': mock.sentinel.weight,
@@ -185,50 +194,17 @@ def expected_sub_config_with_envoy_listeners(expected_sub_config):
     return expected_sub_config
 
 
-def test_get_envoy_listeners():
-    expected_envoy_listeners = {
-        1234: 54321,
-    }
-    mock_envoy_admin_listeners_return_value = {
-        'listener_statuses': [
-            {
-                'name': 'test_service.main.1234.ingress_listener',
-                'local_address': {
-                    'socket_address': {
-                        'address': '0.0.0.0',
-                        'port_value': 54321,
-                    },
-                },
-            },
-        ],
-    }
-    with mock.patch(
-        'nerve_tools.configure_nerve._get_envoy_listeners_from_admin',
-        return_value=mock_envoy_admin_listeners_return_value,
-    ):
-        assert configure_nerve.get_envoy_listeners(123) == \
-            expected_envoy_listeners
-
-
-def test_unsuccessful_get_envoy_listeners():
-    with mock.patch(
-        'nerve_tools.configure_nerve.requests.get',
-        side_effect=Exception,
-    ):
-        assert configure_nerve.get_envoy_listeners(123) == {}
-
-
 def test_generate_subconfiguration(expected_sub_config):
-    with mock.patch(
+    with patch(
         'nerve_tools.configure_nerve.get_current_location',
         side_effect=get_current_location
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.convert_location_type',
         side_effect=convert_location_type
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_named_zookeeper_topology',
         side_effect=get_named_zookeeper_topology
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_labels_by_service_and_port',
         side_effect=get_labels_by_service_and_port
     ):
@@ -265,16 +241,16 @@ def test_generate_subconfiguration(expected_sub_config):
 
 
 def test_generate_subconfiguration_k8s(expected_sub_config):
-    with mock.patch(
+    with patch(
         'nerve_tools.configure_nerve.get_current_location',
         side_effect=get_current_location
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.convert_location_type',
         side_effect=convert_location_type
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_named_zookeeper_topology',
         side_effect=get_named_zookeeper_topology
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_labels_by_service_and_port',
         side_effect=get_labels_by_service_and_port
     ):
@@ -320,21 +296,28 @@ def test_generate_subconfiguration_k8s(expected_sub_config):
     assert new_expected_sub_config == actual_config
 
 
-def test_generate_subconfiguration_with_envoy_listeners(expected_sub_config_with_envoy_listeners):
-    with mock.patch(
+def test_generate_subconfiguration_with_envoy_ingress_listeners(
+    expected_sub_config_with_envoy_ingress_listeners
+):
+    with patch(
         'nerve_tools.configure_nerve.get_current_location',
         side_effect=get_current_location
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.convert_location_type',
         side_effect=convert_location_type
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_named_zookeeper_topology',
         side_effect=get_named_zookeeper_topology
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_labels_by_service_and_port',
         side_effect=get_labels_by_service_and_port
+    ), patch(
+        'nerve_tools.configure_nerve.get_ip_address',
+        return_value='host_ip',
+    ), patch(
+        'nerve_tools.configure_nerve.get_ip_address',
+        return_value='host_ip',
     ):
-
         mock_service_info = {
             'port': 1234,
             'routes': [('remote_location', 'local_location')],
@@ -356,7 +339,7 @@ def test_generate_subconfiguration_with_envoy_listeners(expected_sub_config_with
             'extra_healthcheck_headers': {'Host': 'test_service'},
         })
 
-        actual_config = configure_nerve.generate_subconfiguration(
+        actual_config = generate_subconfiguration(
             service_name='test_service',
             service_info=mock_service_info,
             ip_address='ip_address',
@@ -369,7 +352,60 @@ def test_generate_subconfiguration_with_envoy_listeners(expected_sub_config_with
             envoy_service_info=mock_envoy_service_info,
         )
 
-    assert expected_sub_config_with_envoy_listeners == actual_config
+    assert expected_sub_config_with_envoy_ingress_listeners == actual_config
+
+
+# XXXXXXXXXXXXXXXXXXXXX
+# def test_generate_subconfiguration_with_full_mesh_envoy_listeners(expected_sub_config_with_envoy_listeners):
+#     with patch(
+#         'nerve_tools.configure_nerve.get_current_location',
+#         side_effect=get_current_location
+#     ), patch(
+#         'nerve_tools.configure_nerve.convert_location_type',
+#         side_effect=convert_location_type
+#     ), patch(
+#         'nerve_tools.configure_nerve.get_named_zookeeper_topology',
+#         side_effect=get_named_zookeeper_topology
+#     ), patch(
+#         'nerve_tools.configure_nerve.get_labels_by_service_and_port',
+#         side_effect=get_labels_by_service_and_port
+#     ):
+
+#         mock_service_info = {
+#             'port': 1234,
+#             'routes': [('remote_location', 'local_location')],
+#             'healthcheck_timeout_s': 2.0,
+#             'healthcheck_mode': 'http',
+#             'healthcheck_port': 1234,
+#             'advertise': ['region', 'superregion'],
+#             'extra_advertise': [
+#                 ('habitat:my_habitat', 'region:another_region'),
+#                 ('habitat:your_habitat', 'region:another_region'),  # Ignored
+#             ],
+#             'deploy_group': 'prod.canary',
+#             'paasta_instance': 'canary',
+#         }
+#         mock_envoy_service_info = copy.deepcopy(mock_service_info)
+#         mock_envoy_service_info.update({
+#             'port': 35000,
+#             'healthcheck_port': 35000,
+#             'extra_healthcheck_headers': {'Host': 'test_service'},
+#         })
+
+#         actual_config = configure_nerve.generate_subconfiguration(
+#             service_name='test_service',
+#             service_info=mock_service_info,
+#             ip_address='ip_address',
+#             hacheck_port=6666,
+#             weight=mock.sentinel.weight,
+#             zk_topology_dir='/fake/path',
+#             zk_location_type='superregion',
+#             zk_cluster_type='infrastructure',
+#             labels_dir='/dev/null',
+#             envoy_service_info=mock_envoy_service_info,
+#         )
+
+#     assert expected_sub_config_with_envoy_listeners == actual_config
 
 
 def test_generate_configuration_paasta_service():
@@ -381,13 +417,13 @@ def test_generate_configuration_paasta_service():
         'heartbeat_path': 'test'
     }
 
-    with mock.patch(
+    with patch(
         'nerve_tools.configure_nerve.get_ip_address',
         return_value='ip_address'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_hostname',
         return_value='my_host'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.generate_subconfiguration',
         return_value={'foo': 17}
     ) as mock_generate_subconfiguration:
@@ -412,7 +448,7 @@ def test_generate_configuration_paasta_service():
             zk_location_type='fake_zk_location_type',
             zk_cluster_type='fake_cluster_type',
             labels_dir='/dev/null',
-            envoy_listeners={},
+            envoy_ingress_listeners={},
         )
 
         mock_generate_subconfiguration.assert_called_once_with(
@@ -440,13 +476,13 @@ def test_generate_configuration_paasta_service_with_envoy_listeners():
         'heartbeat_path': 'test'
     }
 
-    with mock.patch(
+    with patch(
         'nerve_tools.configure_nerve.get_ip_address',
         return_value='ip_address'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_hostname',
         return_value='my_host'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.generate_subconfiguration',
         return_value={'foo': 17}
     ) as mock_generate_subconfiguration:
@@ -458,7 +494,7 @@ def test_generate_configuration_paasta_service_with_envoy_listeners():
             'extra_advertise': [('habitat:my_habitat', 'region:another_region')],
         }
 
-        envoy_listeners = {1234: 35001}
+        envoy_ingress_listeners = {1234: 35001}
         mock_envoy_service_main_info = copy.deepcopy(mock_service_info)
         mock_envoy_service_main_info.update({
             'port': 35001,
@@ -491,11 +527,11 @@ def test_generate_configuration_paasta_service_with_envoy_listeners():
             zk_location_type='fake_zk_location_type',
             zk_cluster_type='fake_cluster_type',
             labels_dir='/dev/null',
-            envoy_listeners=envoy_listeners,
+            envoy_ingress_listeners=envoy_ingress_listeners,
         )
 
         mock_generate_subconfiguration.assert_has_calls([
-            mock.call(
+            call(
                 service_name='test_service.main',
                 service_info=mock_service_info,
                 ip_address='ip_address',
@@ -507,7 +543,7 @@ def test_generate_configuration_paasta_service_with_envoy_listeners():
                 labels_dir='/dev/null',
                 envoy_service_info=mock_envoy_service_main_info,
             ),
-            mock.call(
+            call(
                 service_name='test_service.alt',
                 service_info=mock_service_info,
                 ip_address='ip_address',
@@ -533,13 +569,13 @@ def test_generate_configuration_healthcheck_port():
         'heartbeat_path': 'test'
     }
 
-    with mock.patch(
+    with patch(
         'nerve_tools.configure_nerve.get_ip_address',
         return_value='ip_address'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_hostname',
         return_value='my_host'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.generate_subconfiguration',
         return_value={'foo': 17}
     ) as mock_generate_subconfiguration:
@@ -566,7 +602,7 @@ def test_generate_configuration_healthcheck_port():
             zk_location_type='fake_zk_location_type',
             zk_cluster_type='fake_cluster_type',
             labels_dir='/dev/null',
-            envoy_listeners={},
+            envoy_ingress_listeners={},
         )
 
         mock_generate_subconfiguration.assert_called_once_with(
@@ -594,13 +630,13 @@ def test_generate_configuration_healthcheck_mode():
         'heartbeat_path': 'test'
     }
 
-    with mock.patch(
+    with patch(
         'nerve_tools.configure_nerve.get_ip_address',
         return_value='ip_address'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_hostname',
         return_value='my_host'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.generate_subconfiguration',
         return_value={'foo': 17}
     ) as mock_generate_subconfiguration:
@@ -628,7 +664,7 @@ def test_generate_configuration_healthcheck_mode():
             zk_location_type='fake_zk_location_type',
             zk_cluster_type='fake_cluster_type',
             labels_dir='/dev/null',
-            envoy_listeners={},
+            envoy_ingress_listeners={},
         )
 
         mock_generate_subconfiguration.assert_called_once_with(
@@ -648,10 +684,10 @@ def test_generate_configuration_healthcheck_mode():
 
 
 def test_generate_configuration_empty():
-    with mock.patch(
+    with patch(
         'nerve_tools.configure_nerve.get_ip_address',
         return_value='ip_address'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_hostname',
         return_value='my_host'
     ):
@@ -666,46 +702,46 @@ def test_generate_configuration_empty():
             zk_location_type='fake_zk_location_type',
             zk_cluster_type='fake_cluster_type',
             labels_dir='/dev/null',
-            envoy_listeners={},
+            envoy_ingress_listeners={},
         )
         assert configuration == {'instance_id': 'my_host', 'services': {}, 'heartbeat_path': ''}
 
 
 @contextmanager
 def setup_mocks_for_main():
-    mock_sys = mock.MagicMock()
-    mock_file_cmp = mock.Mock()
-    mock_move = mock.Mock()
-    mock_subprocess_call = mock.Mock()
-    mock_subprocess_check_call = mock.Mock()
-    mock_sleep = mock.Mock()
-    mock_file_not_modified = mock.Mock(return_value=False)
+    mock_sys = MagicMock()
+    mock_file_cmp = Mock()
+    mock_move = Mock()
+    mock_subprocess_call = Mock()
+    mock_subprocess_check_call = Mock()
+    mock_sleep = Mock()
+    mock_file_not_modified = Mock(return_value=False)
 
-    with mock.patch.object(
+    with patch.object(
         sys, 'argv', ['configure-nerve']
-    ) as mock_sys, mock.patch(
+    ) as mock_sys, patch(
         'nerve_tools.configure_nerve.get_marathon_services_running_here_for_nerve'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.get_paasta_native_services_running_here_for_nerve'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.generate_configuration'
-    ), mock.patch(
+    ), patch(
         'nerve_tools.configure_nerve.open', create=True
-    ), mock.patch(
+    ), patch(
         'json.dump'
-    ), mock.patch(
+    ), patch(
         'os.chmod'
-    ), mock.patch(
+    ), patch(
         'filecmp.cmp'
-    ) as mock_file_cmp, mock.patch(
+    ) as mock_file_cmp, patch(
         'shutil.move'
-    ) as mock_move, mock.patch(
+    ) as mock_move, patch(
         'subprocess.call'
-    ) as mock_subprocess_call, mock.patch(
+    ) as mock_subprocess_call, patch(
         'subprocess.check_call'
-    ) as mock_subprocess_check_call, mock.patch(
+    ) as mock_subprocess_check_call, patch(
         'time.sleep'
-    ) as mock_sleep, mock.patch(
+    ) as mock_sleep, patch(
         'nerve_tools.configure_nerve.file_not_modified_since', return_value=False
     ) as mock_file_not_modified:
         mocks = (
@@ -718,11 +754,11 @@ def setup_mocks_for_main():
 def test_file_not_modified_since():
     fake_threshold = 10
     fake_path = '/somepath'
-    with mock.patch(
+    with patch(
         'time.time'
-    ) as mock_time, mock.patch(
+    ) as mock_time, patch(
         'os.path.isfile', return_value=True
-    ), mock.patch(
+    ), patch(
         'os.path.getmtime',
     ) as mock_getmtime:
 
@@ -740,17 +776,17 @@ def test_nerve_restarted_when_config_files_differ():
         mock_file_cmp.return_value = False
         configure_nerve.main()
 
-        expected_move = mock.call('/etc/nerve/nerve.conf.json.tmp', '/etc/nerve/nerve.conf.json')
+        expected_move = call('/etc/nerve/nerve.conf.json.tmp', '/etc/nerve/nerve.conf.json')
         assert mock_move.call_args_list == [expected_move]
 
         expected_subprocess_calls = (
-            mock.call(['service', 'nerve-backup', 'start']),
-            mock.call(['service', 'nerve-backup', 'stop']),
+            call(['service', 'nerve-backup', 'start']),
+            call(['service', 'nerve-backup', 'stop']),
         )
         expected_subprocess_check_calls = (
-            mock.call(['service', 'nerve', 'start']),
-            mock.call(['service', 'nerve', 'stop']),
-            mock.call(['/usr/bin/nerve', '-c', '/etc/nerve/nerve.conf.json.tmp', '-k'])
+            call(['service', 'nerve', 'start']),
+            call(['service', 'nerve', 'stop']),
+            call(['/usr/bin/nerve', '-c', '/etc/nerve/nerve.conf.json.tmp', '-k'])
         )
 
         actual_subprocess_calls = mock_subprocess_call.call_args_list
@@ -777,11 +813,11 @@ def test_nerve_not_restarted_when_configs_files_are_identical():
         mock_file_cmp.return_value = True
         configure_nerve.main()
 
-        expected_move = mock.call('/etc/nerve/nerve.conf.json.tmp', '/etc/nerve/nerve.conf.json')
+        expected_move = call('/etc/nerve/nerve.conf.json.tmp', '/etc/nerve/nerve.conf.json')
         assert mock_move.call_args_list == [expected_move]
 
         expected_subprocess_check_calls = [
-            mock.call(['/usr/bin/nerve', '-c', '/etc/nerve/nerve.conf.json.tmp', '-k'])
+            call(['/usr/bin/nerve', '-c', '/etc/nerve/nerve.conf.json.tmp', '-k'])
         ]
 
         actual_subprocess_calls = mock_subprocess_call.call_args_list
@@ -802,17 +838,17 @@ def test_nerve_restarted_when_heartbeat_file_stale():
         mock_file_not_modified.return_value = True
         configure_nerve.main()
 
-        expected_move = mock.call('/etc/nerve/nerve.conf.json.tmp', '/etc/nerve/nerve.conf.json')
+        expected_move = call('/etc/nerve/nerve.conf.json.tmp', '/etc/nerve/nerve.conf.json')
         assert mock_move.call_args_list == [expected_move]
 
         expected_subprocess_calls = (
-            mock.call(['service', 'nerve-backup', 'start']),
-            mock.call(['service', 'nerve-backup', 'stop']),
+            call(['service', 'nerve-backup', 'start']),
+            call(['service', 'nerve-backup', 'stop']),
         )
         expected_subprocess_check_calls = (
-            mock.call(['service', 'nerve', 'start']),
-            mock.call(['service', 'nerve', 'stop']),
-            mock.call(['/usr/bin/nerve', '-c', '/etc/nerve/nerve.conf.json.tmp', '-k'])
+            call(['service', 'nerve', 'start']),
+            call(['service', 'nerve', 'stop']),
+            call(['/usr/bin/nerve', '-c', '/etc/nerve/nerve.conf.json.tmp', '-k'])
         )
 
         actual_subprocess_calls = mock_subprocess_call.call_args_list
@@ -839,11 +875,11 @@ def test_nerve_not_restarted_when_heartbeat_file_valid():
         mock_file_cmp.return_value = True
         configure_nerve.main()
 
-        expected_move = mock.call('/etc/nerve/nerve.conf.json.tmp', '/etc/nerve/nerve.conf.json')
+        expected_move = call('/etc/nerve/nerve.conf.json.tmp', '/etc/nerve/nerve.conf.json')
         assert mock_move.call_args_list == [expected_move]
 
         expected_subprocess_check_calls = [
-            mock.call(['/usr/bin/nerve', '-c', '/etc/nerve/nerve.conf.json.tmp', '-k'])
+            call(['/usr/bin/nerve', '-c', '/etc/nerve/nerve.conf.json.tmp', '-k'])
         ]
 
         actual_subprocess_calls = mock_subprocess_call.call_args_list
